Moved doc_writer_sample back to samples dir
