+++ hydra_python_core/doc_maker.py
@@ -2,7 +2,7 @@
 create a HydraDoc object for it'''
 import re
 import json
-from hydra_python_core.doc_writer_sample import api_doc as sample_document
+from samples.doc_writer_sample import api_doc as sample_document
 from hydra_python_core.doc_writer import HydraDoc, HydraClass, HydraClassProp, HydraClassOp
 from hydra_python_core.doc_writer import HydraStatus
 from typing import Any, Dict, Match, Optional, Tuple, Union
